backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
httpx>=0.27.0
python-telegram-bot>=21.0
aiosqlite>=0.20.0
curl_cffi>=0.14.0
